Rename cursor/subcolumn digits to cells
Cells are a more accurate name for a portion of a subcolumn
that the cursor can occupy at once.
For example, the note subcolumn is 1 cell wide,
meaning the cursor occupies the whole subcolumn (all 3 characters).
Whereas the instrument subcolumn is 2 cells wide,
meaning the cursor can edit each of the 2 digits independently.

In fact, "digit" is a misnomer.
Effect subcolumns are 3-4 cells wide, but only two cells are digits
(1-2 cells are effect name characters).

The subcolumn is the smallest unit of selection and copy-paste,
and the cell is the smallest unit of cursor positioning.