Make synth only process commands (stop/seek) on tick begin
RegisterWriteQueue does not currently allow adding new events
until the previous tick's events have been cleared
at the start of a new tick.

Previously, "stop playback" commands would be processed mid-tick.
If a "stop playback" command writes to the register queue,
it would trip an assertion.
In practice, the upcoming DPCM driver is the first driver to do this.

So only process commands (including "stop playback")
at the beginning of a tick.