Ensure all global variables have the static keyword
-Wmissing-[variable-]declarations warns when you declare a global
function/variable without a forward declaration. This is usually a sign
that you defined it in a .cpp file but didn't declare it in a header,
making it inaccessible to other TUs (so it should be marked static).
Worse yet, it could be *defined* in a header without `inline`, which
would trigger a ODR violation if you included the header twice.

Unfortunately these flags don't function properly on either GCC or
Clang. -Wmissing-declarations fails to catch anything on Clang, and
-Wmissing-variable-declarations is unsupported by GCC. Yay.