Use QWidget::setWindowModified() to control window title
This calls QWidget::setWindowFilePath() and setWindowModified(), which
help produce a more native appearance for document windows and unsaved
changes on Mac.

I don't call QGuiApplication::setApplicationDisplayName(). It's intended
to append an application name to all windows' titles (like
"file.etm - ExoTracker"). In practice it's a disaster:

- It vanishes if your window title ends in ExoTracker (for example if you
  open a file or instrument named ExoTracker).
- Your program doesn't know whether the title is shown.
- The separator differs between Windows and Linux. Your program doesn't
  know what separator is used, and can't control it either.
- If the suffix makes some dialogs' titles too long, you can't turn it
  off on a per-window basis.

As a result, on Linux you can get ugly dialog titles like
"Instrument 00 - Name – ExoTracker" (a hyphen followed by an en dash),
with no way to turn it off.

Instead, only add the app name (with a hyphen) to the main window's
title. One downside is that the title doesn't appear on any dialogs at
all. Another downside is that the title shouldn't be present on Mac, but
I can add a check to remove it if necessary.