On "save as" failure, prompt for a different path
This is different from other apps which keep the old path upon a failed
"save as". However, I think ExoTracker's new behavior may protect users
from losing data: in other apps, if a user edits a document and attempts
to "save as" a new name, but it fails to save (due to permissions or a
read-only drive) and the user doesn't notice the error message, they may
subsequently press Ctrl+S and overwrite the original file by mistake.
ExoTracker instead forces the user to press Cancel if they intend to
keep the old path.