Rewrite timing model from frames/beat fractions to flat ticks
This operates closer to AddMusicK and modern DAWs, making SPC export
through AMK/AMFR more feasible. It eliminates the complexity of beat
fractions, and the redundancy of frames and blocks by removing blocks.

However this makes the timing model less suitable for consoles without a
timer faster than vblank and a fast or dedicated CPU.